requirements.txt
requests
python-dotenv
openai
Flask
pillow